Include dc.json in SIP and remove once loaded
This commit solves the same problem that #676 was trying to solve -
ensure that the `dc.json` file is not included in the AIP. However, by
not including the file in the SIP, #676 also caused `load_dublin_core`
to fail trying to load metadata from that file. This commit reverts the
previous change and instead it removes the file once `load_dublin_core`
is done loading the metadata into the database.

Connects to https://github.com/archivematica/Issues/issues/140.